refactor: handler->api
