+++ api/gacha.go
@@ -1,4 +1,4 @@
-package handler
+package api
 
 import (
 	"bytes"
